Selectable log handler and verbosity (#105)
* Use error_log() for logging

The Monolog handler ErrorLogHandler is based on error_log().
The log is sent through PHP APIs. See the PHP configuration
"error_log" parameter.

* Avoid double timestamps in FPM log

* Support both error_log and stream

The "logfile" .conf parameter decides the message destination. If 
present and not empty a StreamHandler is used, otherwise an error_log 
handler is used.

The new LoggerFactory class instantiates and configures the proper 
Monolog objects.

nethesis/dev#5773